requirements.txt
python-dotenv==1.x
selenium==4.x
pymongo==4.x
flask==2.x
